Improve JaxRsReader to be able to use method.getGenericReturnType
Now the response is Type instead of Class, and this type is used to call ModelConverters, so model converter can handle generics responses too. My actual intentions was to manage responses of type rx.Observable<T> or rx.Single<T> which I map to array of T and T respectively.